hibernate #1 implements

+++ src/main/java/com/game/repository/IPlayerRepository.java
@@ -8,7 +8,7 @@ import java.util.Optional;
 public interface IPlayerRepository {
     List<Player> getAll(int pageNumber, int pageSize);
 
-    int getAllCount();
+    long getAllCount();
 
     Player save(Player player);
